+++ backend/users/views/authentication.py
@@ -83,7 +83,7 @@ class AuthenticationViewSet(GenericViewSet):
         try:
             serializer.is_valid(raise_exception=True)
         except AuthenticationFailed:
-            return Response({"message": "wrong username password"}, status=status.HTTP_401_UNAUTHORIZED)
+            return Response({"detail": "wrong username password"}, status=status.HTTP_401_UNAUTHORIZED)
         except TokenError as e:
             raise InvalidToken(e.args[0])
 
@@ -104,7 +104,7 @@ class AuthenticationViewSet(GenericViewSet):
         try:
             serializer.is_valid(raise_exception=True)
         except AuthenticationFailed:
-            return Response({"message": "wrong username password"}, status=status.HTTP_401_UNAUTHORIZED)
+            return Response({"detail": "wrong username password"}, status=status.HTTP_401_UNAUTHORIZED)
         except TokenError as e:
             raise InvalidToken(e.args[0])